+++ CHANGELOG.md
@@ -1,3 +1,214 @@
+## [6.0.1](https://github.com/adobe/helix-pages/compare/v6.0.0...v6.0.1) (2021-05-08)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-deploy to v4.5.0 ([fa901fb](https://github.com/adobe/helix-pages/commit/fa901fbd1ae4ecfe89ce1c724371894540babeb2))
+* **deps:** update dependency @adobe/helix-universal to v1.5.0 ([6a771bd](https://github.com/adobe/helix-pages/commit/6a771bde7c6bff71d3abce96c5b634471cea919d))
+
+# [6.0.0](https://github.com/adobe/helix-pages/compare/v5.0.4...v6.0.0) (2021-05-06)
+
+
+* Merge pull request #662 from adobe/breaking-202103 ([f00f452](https://github.com/adobe/helix-pages/commit/f00f452bc8a52a97e296fbb40b373f9a537382da)), closes [#662](https://github.com/adobe/helix-pages/issues/662) [#657](https://github.com/adobe/helix-pages/issues/657) [#638](https://github.com/adobe/helix-pages/issues/638) [#728](https://github.com/adobe/helix-pages/issues/728) [#729](https://github.com/adobe/helix-pages/issues/729) [#745](https://github.com/adobe/helix-pages/issues/745) [#786](https://github.com/adobe/helix-pages/issues/786) [#796](https://github.com/adobe/helix-pages/issues/796) [#806](https://github.com/adobe/helix-pages/issues/806) [#748](https://github.com/adobe/helix-pages/issues/748) [adobe/helix-pipeline#1005](https://github.com/adobe/helix-pipeline/issues/1005)
+
+
+### Bug Fixes
+
+* **breaking:** make branch unsticky ([0478b9a](https://github.com/adobe/helix-pages/commit/0478b9a3e6912dee3d1e34340f8ba8d60601b879))
+* **breaking:** revert metadata from spreadsheet ([a59601b](https://github.com/adobe/helix-pages/commit/a59601b881a910bcc0ae5b1eef8cd721b9240383)), closes [#801](https://github.com/adobe/helix-pages/issues/801)
+* **config:** add breaking strain and version file ([9513746](https://github.com/adobe/helix-pages/commit/9513746eb65f677fa7b07ef5bb2a10ca61823a26))
+* **deps:** update adobe fixes ([#661](https://github.com/adobe/helix-pages/issues/661)) ([161bf09](https://github.com/adobe/helix-pages/commit/161bf09547528ce220355829597fdac918a15e54))
+* **deps:** update adobe fixes ([#881](https://github.com/adobe/helix-pages/issues/881)) ([09d5c77](https://github.com/adobe/helix-pages/commit/09d5c77ca354e2924b83f6b717808b736c96712f))
+* **deps:** use universal-logger ([c3ba89a](https://github.com/adobe/helix-pages/commit/c3ba89a683fbffb1a93af76b8b6a717d4ee25c0a))
+* **html:** fix sections for also plain.html ([65a78de](https://github.com/adobe/helix-pages/commit/65a78def95da7fc1d1f43ff25a25f5888db8d3cd)), closes [#748](https://github.com/adobe/helix-pages/issues/748)
+* **meta:** hero image is not correct ([#775](https://github.com/adobe/helix-pages/issues/775)) ([b510376](https://github.com/adobe/helix-pages/commit/b510376f7fb607de1230d3b0054efbec4c874808)), closes [#745](https://github.com/adobe/helix-pages/issues/745)
+* **pageblocks:** avoid errors if empty table header ([40b57be](https://github.com/adobe/helix-pages/commit/40b57be219b16f0d61f805d2e43d95ddf5e3c050))
+* **pageblocks:** remove classes from cells ([04fff0c](https://github.com/adobe/helix-pages/commit/04fff0cc9c62e935b3bc9cd30c4f3793cc945af4))
+* **picture:** use backward compatible format ([0491276](https://github.com/adobe/helix-pages/commit/049127683ff30188ccae73078ac8420b96ed0c0b))
+* **pre:** adjust image optimization ([#828](https://github.com/adobe/helix-pages/issues/828)) ([743911b](https://github.com/adobe/helix-pages/commit/743911b8e75cf772249052aca49d360b0be808e6)), closes [#657](https://github.com/adobe/helix-pages/issues/657)
+* **seo:** canonical url enforces html extension ([5f121cd](https://github.com/adobe/helix-pages/commit/5f121cdd7ff34baedfe7776cc19afc266449ef80))
+* **seo:** detect meta title inside block ([6a25bef](https://github.com/adobe/helix-pages/commit/6a25bef371d2e81e51ad8b33e73f4051e470f7e6))
+* **seo:** meta title is not document title ([62b8533](https://github.com/adobe/helix-pages/commit/62b85336d73cb628142761fff6060864d05d4ea9))
+* **sidekick:** endless update loop ([a3ca8f7](https://github.com/adobe/helix-pages/commit/a3ca8f7619dc30772f924aa979429487eefafae1))
+
+
+### Features
+
+* **blob:** adjust to rootless image paths ([daaf3b3](https://github.com/adobe/helix-pages/commit/daaf3b33cc2238dab89841659d20a4030268bb40))
+* **blob:** adjust to rootless image paths ([91ad284](https://github.com/adobe/helix-pages/commit/91ad2844c06944f9d2c232067f2bf2922895b1ee))
+* **blob:** adjust to rootless image paths ([ba541cd](https://github.com/adobe/helix-pages/commit/ba541cd93f90f1746352bf0fbe3251817f2c3856))
+* **blobs:** adjust to rootless image paths ([6dce2ec](https://github.com/adobe/helix-pages/commit/6dce2ec5936adac58abbdc566b223560ed008476))
+* **html:** all img tags ([00c3bbf](https://github.com/adobe/helix-pages/commit/00c3bbfdba7626ac4aea268b7595b15faa26e90f))
+* **html:** responsive images with <picture> ([#657](https://github.com/adobe/helix-pages/issues/657)) ([4e81f25](https://github.com/adobe/helix-pages/commit/4e81f25ab81df15b3972500a3520307eeb0ba43d))
+* **html:** responsive images with <picture> ([#657](https://github.com/adobe/helix-pages/issues/657)) ([ee700b5](https://github.com/adobe/helix-pages/commit/ee700b5adac46b88b547cbcd539ffa86d6abfc8f))
+* **html:** responsive images with <picture> ([#657](https://github.com/adobe/helix-pages/issues/657)) ([31bcc77](https://github.com/adobe/helix-pages/commit/31bcc77a61d543e35ca286364e5e68d6c779ddee))
+* **meta:** add support for metadata with array of values (multiple p or ol/ul) ([4f03461](https://github.com/adobe/helix-pages/commit/4f03461fc0ac525e59556fb2e015ae3a100418a3))
+* **metadata:** support line-separated lists and custom metadata ([688e9c0](https://github.com/adobe/helix-pages/commit/688e9c0fa4d41a117a6c43c7687fdf81fe19f6cb))
+* **render:** wrap all sections in 2 divs ([85990f9](https://github.com/adobe/helix-pages/commit/85990f9983fc0d5731112fb87a8e096209d6c171)), closes [#640](https://github.com/adobe/helix-pages/issues/640)
+* **seo:** metadata block handling ([fbfa503](https://github.com/adobe/helix-pages/commit/fbfa503c02e424b47985f5114cfa7de004277e18))
+* **seo:** metadata from spreadsheet ([#801](https://github.com/adobe/helix-pages/issues/801)) ([cd6b9db](https://github.com/adobe/helix-pages/commit/cd6b9db1d47c2b98af4229ade61a5b4a8b13bdb5))
+* **seo:** use meta image from block ([5ffa59e](https://github.com/adobe/helix-pages/commit/5ffa59e15310949cfe81b7508689af9abd37b996))
+* inplement server side blocks handling ([#702](https://github.com/adobe/helix-pages/issues/702)) ([73c1a0a](https://github.com/adobe/helix-pages/commit/73c1a0a2bcfce0783c072d0a4094b9ec8c06adae)), closes [#638](https://github.com/adobe/helix-pages/issues/638)
+
+
+### BREAKING CHANGES
+
+* march 2021
+
+## [5.0.4](https://github.com/adobe/helix-pages/compare/v5.0.3...v5.0.4) (2021-04-28)
+
+
+### Bug Fixes
+
+* **sidekick:** auto-heal ([8d0e1e4](https://github.com/adobe/helix-pages/commit/8d0e1e4561930086244ff6f402a23fd8b78361e1))
+* **sidekick:** byocdn flag always set on update ([7622deb](https://github.com/adobe/helix-pages/commit/7622debaf2945c4234c11c00f0f78693e452235a))
+* **sidekick:** trigger update dialog on affected configs ([d113272](https://github.com/adobe/helix-pages/commit/d1132729feff70d1f77aa6b045250cfbff9f2ce4))
+
+## [5.0.3](https://github.com/adobe/helix-pages/compare/v5.0.2...v5.0.3) (2021-04-23)
+
+
+### Bug Fixes
+
+* **vcl:** reduce workspace usage ([3ff5f21](https://github.com/adobe/helix-pages/commit/3ff5f21dd6b3d98be6da5f8cb23d370853910d67))
+
+## [5.0.2](https://github.com/adobe/helix-pages/compare/v5.0.1...v5.0.2) (2021-04-22)
+
+
+### Bug Fixes
+
+* trigger publish ([d946a05](https://github.com/adobe/helix-pages/commit/d946a05b23b81a4c87d806c90cd3b89f86f8604b))
+
+## [5.0.1](https://github.com/adobe/helix-pages/compare/v5.0.0...v5.0.1) (2021-04-22)
+
+
+### Bug Fixes
+
+* **sidekick:** share url missing byocdn flag ([a9099a7](https://github.com/adobe/helix-pages/commit/a9099a76cfa08b812c9b12316c63cbed260b9412))
+
+# [5.0.0](https://github.com/adobe/helix-pages/compare/v4.28.0...v5.0.0) (2021-04-21)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([d952bec](https://github.com/adobe/helix-pages/commit/d952becbd75486ed15b88a0b8dc064b2336f089d))
+
+
+### Features
+
+* **sidekick:** switch default branch to main in bookmarklet generator ([19357fe](https://github.com/adobe/helix-pages/commit/19357fe9fee593dd59945c97288b45c7703be96e))
+
+
+### BREAKING CHANGES
+
+* **sidekick:** projects still using master as default need to specify it as part of the gh url, e.g. https://github/owner/repo/tree/master
+
+# [4.28.0](https://github.com/adobe/helix-pages/compare/v4.27.2...v4.28.0) (2021-04-21)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#874](https://github.com/adobe/helix-pages/issues/874)) ([86329a4](https://github.com/adobe/helix-pages/commit/86329a4b496e91ce2345b45e30f3034e6222466d))
+* accelerate loading plugins in editor ([5554c40](https://github.com/adobe/helix-pages/commit/5554c406b162cc882e955d302dfc6835c8853b0d))
+* **sidekick:** innerHost is hlx.live ([943eed0](https://github.com/adobe/helix-pages/commit/943eed0902247fa662ebf37a20cd1271170d421e))
+
+
+### Features
+
+* **sidekick:** always reload context when activated ([6122e94](https://github.com/adobe/helix-pages/commit/6122e9428c77e0636a9a942d72ae389bf42c44b3))
+* **sidekick:** load sidekick from same host as bookmarklet generator ([b7f3893](https://github.com/adobe/helix-pages/commit/b7f38938996179a521e55415921541919f98278d))
+* **sidekick:** middle click and meta key support for all plugins ([8bc4db5](https://github.com/adobe/helix-pages/commit/8bc4db50d4249b1c0c52a1824ec15e6d23ec186d))
+* **sidekick:** middle click and meta key support for all plugins ([f5bd6c0](https://github.com/adobe/helix-pages/commit/f5bd6c037dc0328c3d301c1abc1daaa3b4915fee))
+* **sidekick:** middle click and meta key support for all plugins ([af52552](https://github.com/adobe/helix-pages/commit/af52552cb68244c7792b0f8d626ca452c2033a60))
+* **sidekick:** no publish button on byocdn prod host ([20c1a21](https://github.com/adobe/helix-pages/commit/20c1a2136b7534a3f55fe1be6ba69196d9d2b3fc))
+* **sidekick:** use .lnk URLs ([2ab96ab](https://github.com/adobe/helix-pages/commit/2ab96ab71f790eafa5f5043e29b59d2c0e00b1a1))
+* **sidekick:** use HLXPURGE ([1a33116](https://github.com/adobe/helix-pages/commit/1a3311679b75a7e0d3438a207dd028a8b4a20fe0))
+* **sidekick:** write sidekick config on every invocation ([b5f53f6](https://github.com/adobe/helix-pages/commit/b5f53f6cc38f605001157e3b7b799076bac6e78f))
+
+## [4.27.2](https://github.com/adobe/helix-pages/compare/v4.27.1...v4.27.2) (2021-04-20)
+
+
+### Bug Fixes
+
+* **extensions:** allow access to private, acl-restricted content with github token ([ee45ff9](https://github.com/adobe/helix-pages/commit/ee45ff9ff739695c553aba3c9ac8841e90356b73))
+* correct client AS number for Fastly ([72f818e](https://github.com/adobe/helix-pages/commit/72f818e71932265f9f6843e9e1022a22070cde14))
+* **extensions.vcl:** don't apply acl filter to PURGE requests ([63b86e8](https://github.com/adobe/helix-pages/commit/63b86e8f5f59b36babb8398af53d6e8a6290b786))
+
+## [4.27.1](https://github.com/adobe/helix-pages/compare/v4.27.0...v4.27.1) (2021-04-20)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-shared to v7.26.3 ([e4d4e67](https://github.com/adobe/helix-pages/commit/e4d4e67a6b9d643e61810961aaf907524e41945b))
+
+# [4.27.0](https://github.com/adobe/helix-pages/compare/v4.26.1...v4.27.0) (2021-04-19)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-pipeline to v13.9.8 ([#870](https://github.com/adobe/helix-pages/issues/870)) ([b9bdb5d](https://github.com/adobe/helix-pages/commit/b9bdb5d916c9e397289d36419de060ae441a4115))
+
+
+### Features
+
+* **extension.vcl:** support private gh repos and acls ([ddbc7e1](https://github.com/adobe/helix-pages/commit/ddbc7e11cde044d913915f2d41fd05a679e58bfa))
+
+## [4.26.1](https://github.com/adobe/helix-pages/compare/v4.26.0...v4.26.1) (2021-04-19)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#865](https://github.com/adobe/helix-pages/issues/865)) ([757495f](https://github.com/adobe/helix-pages/commit/757495fd8c3b2f101ec49daac3171abd30a81317))
+* **deps:** update external fixes ([#869](https://github.com/adobe/helix-pages/issues/869)) ([d81dd95](https://github.com/adobe/helix-pages/commit/d81dd95b1aafde6577b8b9acfba81f84e8f68ded))
+
+# [4.26.0](https://github.com/adobe/helix-pages/compare/v4.25.6...v4.26.0) (2021-04-16)
+
+
+### Features
+
+* **sidekick:** always include branch in preview host ([5f92801](https://github.com/adobe/helix-pages/commit/5f9280191914612021187cb830e39ce67f5595be))
+* **sidekick:** use main as default branch ([2a4e3e9](https://github.com/adobe/helix-pages/commit/2a4e3e9880e3c40b2d5ee65eba892b8b72679aa9))
+
+## [4.25.6](https://github.com/adobe/helix-pages/compare/v4.25.5...v4.25.6) (2021-04-16)
+
+
+### Bug Fixes
+
+* **deploy:** use google-compatible package version names ([71f0c01](https://github.com/adobe/helix-pages/commit/71f0c01c8cfdff79c0ea70b6605df526cc50ccfb))
+
+
+### Reverts
+
+* Revert "Revert "Merge pull request #840 from adobe/deploy-gooooooooooogle"" ([a2d4260](https://github.com/adobe/helix-pages/commit/a2d42602ef06c375ad42fa57beaba419f673ab33)), closes [#840](https://github.com/adobe/helix-pages/issues/840)
+
+## [4.25.5](https://github.com/adobe/helix-pages/compare/v4.25.4...v4.25.5) (2021-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-pipeline to v13.9.4 ([760dd4c](https://github.com/adobe/helix-pages/commit/760dd4cd98a74bffe51cdf2772d2c2c6783bd0ce))
+
+## [4.25.4](https://github.com/adobe/helix-pages/compare/v4.25.3...v4.25.4) (2021-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-shared to v7.25.1 ([ddbc71c](https://github.com/adobe/helix-pages/commit/ddbc71ccfd816798d08626c387192079e7802816))
+
+## [4.25.3](https://github.com/adobe/helix-pages/compare/v4.25.2...v4.25.3) (2021-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @adobe/helix-pipeline to v13.9.3 ([#862](https://github.com/adobe/helix-pages/issues/862)) ([13fd3c9](https://github.com/adobe/helix-pages/commit/13fd3c98414f6702690d326e1f0216aa7a393ccd))
+
+## [4.25.2](https://github.com/adobe/helix-pages/compare/v4.25.1...v4.25.2) (2021-04-15)
+
+
+### Bug Fixes
+
+* **deps:** update adobe fixes ([#860](https://github.com/adobe/helix-pages/issues/860)) ([4ecf1ef](https://github.com/adobe/helix-pages/commit/4ecf1ef06ca1ba8d5b88d4a6ccb4be789f511c8a))
+* **deps:** update dependency @adobe/helix-shared to v7.25.0 ([#861](https://github.com/adobe/helix-pages/issues/861)) ([85f8368](https://github.com/adobe/helix-pages/commit/85f8368bef6a7d57c9a8e3a74046e61e9f9696be))
+
 ## [4.25.1](https://github.com/adobe/helix-pages/compare/v4.25.0...v4.25.1) (2021-04-13)